# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Verde Blocks",
        "description": "Swagger UI for VerdeBlocks",
        "termsOfService": "http://swagger.io/terms/"
    },
    "host": "52.66.154.45:5000",
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "User Login",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User Login Request",
                        "schema": {
                            "$ref": "#/definitions/LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User SignUp success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "User does not exists",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/setPassword": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Set Password",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Add set password Request",
                        "schema": {
                            "$ref": "#/definitions/SetPasswordResquest"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "schema": {
                            "$ref": "#/definitions/GetAllBuyers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Set password success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/RPP/{RPPId}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get RPP",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "schema": {
                            "$ref": "#/definitions/GetAllBuyers"
                        }
                    },
                    {
                        "name": "RPPId",
                        "in": "path",
                        "description": "RPP id",
                        "schema": {
                            "$ref": "#/definitions/GetAllBuyers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get RPP success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/ISO": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get all ISOs",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "schema": {
                            "$ref": "#/definitions/GetAllBuyers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all ISO success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/zone/{ISO}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get all zones in an ISO",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "schema": {
                            "$ref": "#/definitions/GetAllZones"
                        }
                    },
                    {
                        "name": "ISO",
                        "in": "path",
                        "description": "ISO",
                        "schema": {
                            "$ref": "#/definitions/GetAllZones"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all zones of an ISO success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/timeZone": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get all Time Zones",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "schema": {
                            "$ref": "#/definitions/GetAllBuyers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all time zones success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/states": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get all states for RECs",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "schema": {
                            "$ref": "#/definitions/GetAllBuyers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all ISO success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/RECTypes/{state}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get all RECTypes of a state for RECs",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "schema": {
                            "$ref": "#/definitions/GetAllRECTypes"
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "description": "state",
                        "schema": {
                            "$ref": "#/definitions/GetAllRECTypes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all zones of an ISO success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/years/{RECType}/{state}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get all years of a RECType of a state for RECs",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "schema": {
                            "$ref": "#/definitions/GetAllYears"
                        }
                    },
                    {
                        "name": "RECType",
                        "in": "path",
                        "description": "RECType",
                        "schema": {
                            "$ref": "#/definitions/GetAllYears"
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "description": "state",
                        "schema": {
                            "$ref": "#/definitions/GetAllYears"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all zones of an ISO success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/getSellers": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get all sellers id and company name",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "schema": {
                            "$ref": "#/definitions/GetAllYears"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all sellers success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/{id}/buyer": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "description": "Get all Buyer",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin id",
                        "schema": {
                            "$ref": "#/definitions/GetAllBuyers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all Buyer success response",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty field in request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error response from DB",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "LoginRequest": {
            "properties": {
                "email": {
                    "required": true,
                    "type": "string",
                    "example": "admin@verdeblocks.com"
                },
                "password": {
                    "required": true,
                    "type": "string",
                    "example": "User@123"
                }
            }
        },
        "SetPasswordResquest": {
            "properties": {
                "oldPassword": {
                    "required": true,
                    "type": "string",
                    "example": "Test@123"
                },
                "newPassword": {
                    "required": true,
                    "type": "string",
                    "example": "Abc@123"
                }
            }
        },
...
        "GetAllBuyers": {
            "properties": {
                "id": {
                    "example": 1,
                    "type": "number"
                }
            }
        },
...
        "SuccessResponse": {
            "properties": {
                "message": {
                    "example": "Success message!"
                },
                "success": {
                    "type": "bool",
                    "example": true
                }
            }
        },
        "ErrorResponse": {
            "properties": {
                "error": {
                    "example": "contains error object or message"
                },
                "success": {
                    "type": "bool",
                    "example": false
                }
            }
        }
    }
}